Progress Bar
Each plant has a bounding box above it to detect if the watering can is pouring. This is achieved by the WateringCanDetector and WaterReceiver scripts. The detector goes on the watering can object (specifically it's origin) and the receiver goes on the plant's bounding box. The progress bar attached to the plant's bounding box has "Growth Stage Thresholds" where the plant will advance to it's next stage. For example, a threshold of 0.5 and 1 w/ a growth time of 4 will advance the plant's model to stage 2 at 2 seconds, and to stage 3 at 4 seconds. The script is malleable, meaning animations and other stuffs should be easy to add/replace.