Added popup functionality for in-between thresholds and completed growth stages.
Added fact-popups in the form of images for x thresholds during the growth stage and a final popup w/ a "Cool!" button implemented.

I haven't tested the in-between popups that thoroughly so it might need to be fixed. But the finalpopup should be gucci.

For every growth stage threshold, there can be a popup attached to it. The popup is in the form of an image and UI position. Currently, there are not UI positions because there are no in-between fun fact popups.

When the progress bar is full, a final popup in the form of an image and button is set to active. Yet unlike the growth stage popups, the final fact stays on screen until the cool button is pressed, then dissapears from the screen.

I have linked the corresponding popups w/ each plant, so I'm pretty sure the progress bar stuffs are done :)))